Multiple loops implemented with WP_query
Also created a function that grabs the category ID when it is passed the category slug. This gives us the freedom to change the category name as we see fit while not breaking anything.

The Slug should never be changed anyways since the category URL is dependent on the slug